say my name3

+++ lib/view/product_details.dart
@@ -1,4 +1,3 @@
-import 'package:amf/data/webserviecs/webservice.dart';
 import 'package:amf/imagecachemanager.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -7,7 +6,7 @@ import '../cubit/product_cubit.dart';
 import '../data/models/product.dart';
 
 class ProductDetail extends StatefulWidget {
-  final Product product;
+   final Product product;
 
   const ProductDetail({required this.product, super.key});
 
@@ -22,6 +21,14 @@ class _ProductDetailState extends State<ProductDetail> {
   void initState() {
     super.initState();
     _product = widget.product;
+    BlocProvider.of<ProductCubit>(context).fetchProduct(_product.id!);
+  }
+  @override
+  void setState(VoidCallback fn) {
+    // TODO: implement setState
+    super.setState(fn);
+    BlocProvider.of<ProductCubit>(context).fetchProduct(_product.id!);
+
   }
 
   @override
@@ -31,18 +38,23 @@ class _ProductDetailState extends State<ProductDetail> {
     double height = MediaQuery.of(context).size.height;
 
 
-    return Scaffold(
-      resizeToAvoidBottomInset: false,
-      backgroundColor: const Color.fromRGBO(35, 52, 69, 10),
-      floatingActionButton: FloatingActionButton(
-          child: const Icon(Icons.edit),
-          onPressed: () {
-            Navigator.pushNamed(context, 'edit',arguments: _product ).then((value) => setState((){
-              _product;
+    return SafeArea(
+      child: Scaffold(
+        resizeToAvoidBottomInset: false,
+        backgroundColor: Colors.white,
+        floatingActionButton: FloatingActionButton(
+            child: const Icon(Icons.edit),
+            onPressed: () async{
+            await  Navigator.pushNamed(context, 'edit',arguments: _product ).then((value) => setState((){
+              BlocProvider.of<ProductCubit>(context).fetchProduct(_product.id!);
+
             }));
-          }),
-      body: SafeArea(
-        child: CustomScrollView(
+            }),
+        body: BlocBuilder<ProductCubit, ProductState>(
+  builder: (context, state) {
+    if(state is OneProductAdded){
+      final p = state.product;
+    return CustomScrollView(
           physics: const ClampingScrollPhysics(),
 
           //physics: NeverScrollableScrollPhysics(),
@@ -50,7 +62,7 @@ class _ProductDetailState extends State<ProductDetail> {
             SliverAppBar(
               elevation: 0,
               title: Text(
-                '${_product.title}',
+                '${p.title}',
                 style: const TextStyle(
                     fontWeight: FontWeight.bold, color: Colors.blueGrey),
                 textAlign: TextAlign.start,
@@ -59,7 +71,30 @@ class _ProductDetailState extends State<ProductDetail> {
               actions: [
                 IconButton(
                     onPressed: () async {
-                      await BlocProvider.of<ProductCubit>(context).deletePro(_product.id!, context);
+                      showDialog(
+                        context: context,
+                        builder: (ctx) => AlertDialog(
+                          icon: const Icon(Icons.add_alert),
+                          title: const Text("Delete!"),
+                          content: const Text("Do you want to Delete this product?",style: TextStyle(fontWeight: FontWeight.bold),),
+                          actions: <Widget>[
+                            TextButton(
+                              child: const Text('Undo'),
+                              onPressed: () {
+                                Navigator.pop(context);
+                              },
+                            ),
+                            TextButton(
+                              child: const Text('Delete'),
+                              onPressed: () async{
+                                await BlocProvider.of<ProductCubit>(context).deletePro(_product.id!, context);
+                                Navigator.pop(context);
+                                setState(() {});
+                              },
+                            ),
+                          ],
+                        ),
+                      );
                     },
                     icon: const Icon(
                       Icons.delete_forever_outlined,
@@ -75,7 +110,7 @@ class _ProductDetailState extends State<ProductDetail> {
                 background: Hero(
                   tag: _product.id ?? 0,
                   child: ImageCacheManager().getImage(
-                    '${_product.pictureLink}',
+                    '${p.pictureLink}',
                   ),
                 ),
               ),
@@ -94,7 +129,7 @@ class _ProductDetailState extends State<ProductDetail> {
                         RichText(
                             text: const TextSpan(
                                 text: 'Size : ',
-                                style: TextStyle(
+                                style: TextStyle(color: Colors.black,
                                     fontSize: 25, fontWeight: FontWeight.w500)))
                       ],
                     ),
@@ -107,8 +142,9 @@ class _ProductDetailState extends State<ProductDetail> {
                           alignment: Alignment.centerLeft,
                           child: RichText(
                               text: TextSpan(
-                                  text: '${_product.size}',
+                                  text: '${p.size}',
                                   style: const TextStyle(
+                                      color: Colors.black,
                                       fontSize: 20,
                                       fontWeight: FontWeight.w300))),
                         ),
@@ -122,7 +158,7 @@ class _ProductDetailState extends State<ProductDetail> {
                         RichText(
                             text: const TextSpan(
                                 text: 'Brand :',
-                                style: TextStyle(
+                                style: TextStyle(color: Colors.black,
                                     fontSize: 25, fontWeight: FontWeight.w500)))
                       ],
                     ),
@@ -135,8 +171,8 @@ class _ProductDetailState extends State<ProductDetail> {
                           alignment: Alignment.centerLeft,
                           child: RichText(
                               text: TextSpan(
-                                  text: '${_product.brand}',
-                                  style: const TextStyle(
+                                  text: '${p.brand}',
+                                  style: const TextStyle(color: Colors.black,
                                       fontSize: 20,
                                       fontWeight: FontWeight.w300))),
                         ),
@@ -150,7 +186,7 @@ class _ProductDetailState extends State<ProductDetail> {
                         RichText(
                             text: const TextSpan(
                                 text: 'Color :',
-                                style: TextStyle(
+                                style: TextStyle(color: Colors.black,
                                     fontSize: 25, fontWeight: FontWeight.w500)))
                       ],
                     ),
@@ -164,7 +200,7 @@ class _ProductDetailState extends State<ProductDetail> {
                           child: RichText(
                               text: TextSpan(
                                   text: '${_product.color}',
-                                  style: const TextStyle(
+                                  style: const TextStyle(color: Colors.black,
                                       fontSize: 20,
                                       fontWeight: FontWeight.w300))),
                         ),
@@ -180,7 +216,7 @@ class _ProductDetailState extends State<ProductDetail> {
                           text: const TextSpan(children: [
                             TextSpan(
                               text: 'Additions :  ',
-                              style: TextStyle(
+                              style: TextStyle(color: Colors.black,
                                   fontWeight: FontWeight.w500, fontSize: 25),
                             ),
                           ]),
@@ -197,8 +233,8 @@ class _ProductDetailState extends State<ProductDetail> {
                           child: RichText(
                               textAlign: TextAlign.left,
                               text: TextSpan(
-                                  text: '${_product.addtions}',
-                                  style: const TextStyle(
+                                  text: '${p.addtions}',
+                                  style: const TextStyle(color: Colors.black,
                                       fontSize: 20,
                                       fontWeight: FontWeight.w300))),
                         ),
@@ -209,7 +245,11 @@ class _ProductDetailState extends State<ProductDetail> {
               ),
             )
           ],
-        ),
+        );
+  }
+    return const Center(child: CircularProgressIndicator());
+    },
+),
       ),
     );
   }
